# docs/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"UserSettings": {
				"properties": {
					"id": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"userId": {
						"type": "string"
					},
					"notification": {},
					"preference": {
						"type": "string",
						"enum": [
							"system",
							"dark",
							"light"
						]
					}
				},
				"required": [
					"id",
					"createdAt",
					"updatedAt",
					"userId",
					"notification",
					"preference"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Organiser": {
				"properties": {
					"id": {
						"type": "string"
					},
					"phoneNumber": {
						"type": "string"
					},
					"phoneNumberVerified": {
						"type": "boolean"
					},
					"address": {
						"type": "string",
						"nullable": true
					},
					"logo": {
						"type": "string",
						"nullable": true
					},
					"bio": {
						"type": "string",
						"nullable": true
					},
					"website": {
						"type": "string",
						"nullable": true
					},
					"email": {
						"type": "string"
					},
					"organisationName": {
						"type": "string"
					},
					"userId": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"phoneNumber",
					"phoneNumberVerified",
					"address",
					"logo",
					"bio",
					"website",
					"email",
					"organisationName",
					"userId",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AccountOutput": {
				"properties": {
					"id": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"userId": {
						"type": "string"
					},
					"providerId": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"createdAt",
					"updatedAt",
					"userId",
					"providerId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"OrganizerPayoutAccount": {
				"properties": {
					"id": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"userId": {
						"type": "string"
					},
					"bvn": {
						"type": "string",
						"nullable": true
					},
					"bankName": {
						"type": "string"
					},
					"paystackSubAccountCode": {
						"type": "string"
					},
					"bankAccountName": {
						"type": "string"
					},
					"bankAccountNumber": {
						"type": "string"
					},
					"bankRoutingNumber": {
						"type": "string",
						"nullable": true
					},
					"verified": {
						"type": "boolean",
						"nullable": true
					},
					"schedulePayout": {
						"type": "string",
						"enum": [
							"weekly",
							"monthly"
						]
					}
				},
				"required": [
					"id",
					"createdAt",
					"updatedAt",
					"userId",
					"bvn",
					"bankName",
					"paystackSubAccountCode",
					"bankAccountName",
					"bankAccountNumber",
					"bankRoutingNumber",
					"verified",
					"schedulePayout"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserWithRelationship": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"emailVerified": {
						"type": "boolean"
					},
					"image": {
						"type": "string",
						"nullable": true
					},
					"role": {
						"type": "string",
						"enum": [
							"user",
							"admin",
							"staff",
							"hoster"
						]
					},
					"banned": {
						"type": "boolean",
						"nullable": true
					},
					"customerCode": {
						"type": "string",
						"nullable": true
					},
					"lastLogin": {
						"type": "string",
						"format": "date-time"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"userSettings": {
						"$ref": "#/components/schemas/UserSettings"
					},
					"organiser": {
						"$ref": "#/components/schemas/Organiser"
					},
					"accounts": {
						"items": {
							"$ref": "#/components/schemas/AccountOutput"
						},
						"type": "array"
					},
					"organizerPayoutAccount": {
						"$ref": "#/components/schemas/OrganizerPayoutAccount"
					}
				},
				"required": [
					"id",
					"name",
					"email",
					"emailVerified",
					"image",
					"role",
					"banned",
					"customerCode",
					"lastLogin",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"APIResponse_UserWithRelationship-or-null_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/UserWithRelationship"
							}
						],
						"nullable": true
					},
					"error": {
						"type": "string"
					},
					"message": {}
				},
				"required": [
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"APIResponse_null_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"data": {
						"anyOf": [],
						"nullable": true
					},
					"error": {
						"type": "string"
					},
					"message": {}
				},
				"required": [
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"CreateUserInputBody": {
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"role": {
						"type": "string",
						"enum": [
							"admin",
							"staff",
							"hoster",
							"user"
						]
					},
					"image": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"password",
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"EmailVericationBodyInput": {
				"properties": {
					"token": {
						"type": "string"
					}
				},
				"required": [
					"token"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PaginatedResult_UserWithRelationship_": {
				"description": "Represents a paginated result set with metadata.",
				"properties": {
					"success": {
						"type": "boolean"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/UserWithRelationship"
						},
						"type": "array",
						"description": "Array of paginated items"
					},
					"pageSize": {
						"type": "number",
						"format": "double",
						"description": "Number of items per page"
					},
					"nextCursor": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "string",
								"format": "date-time"
							},
							{
								"type": "number",
								"format": "double"
							}
						],
						"nullable": true,
						"description": "The cursor for the next page (null if no more pages)"
					},
					"hasNextPage": {
						"type": "boolean",
						"description": "Whether there are more pages available"
					}
				},
				"required": [
					"success",
					"data",
					"pageSize",
					"nextCursor",
					"hasNextPage"
				],
				"type": "object",
				"additionalProperties": false
			},
			"APIResponse_PaginatedResult_UserWithRelationship_-or-null_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/PaginatedResult_UserWithRelationship_"
							}
						],
						"nullable": true
					},
					"error": {
						"type": "string"
					},
					"message": {}
				},
				"required": [
					"success"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"bearer": {
				"type": "http",
				"name": "Authorization",
				"scheme": "bearer",
				"in": "header",
				"bearerFormat": "JWT",
				"description": "Enter your bearer token in the format \"Bearer {token}\""
			},
			"tsoa_auth": {
				"type": "oauth2",
				"flows": {
					"implicit": {
						"authorizationUrl": "http://swagger.io/api/oauth/dialog",
						"scopes": {
							"write:pets": "modify things",
							"read:pets": "read things"
						}
					}
				}
			}
		}
	},
	"info": {
		"title": "luminevent",
		"version": "1.2.1",
		"description": "Luminevent is a backend service for managing and organizing events, providing APIs for event creation, user management, and scheduling.",
		"license": {
			"name": "Apache-2.0"
		},
		"contact": {
			"name": "BITS"
		}
	},
	"paths": {
...
		"/auth/verify-email": {
			"post": {
				"operationId": "VerifyEmail",
				"responses": {
					"200": {
						"description": "Email verified successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIResponse_UserWithRelationship-or-null_"
								}
							}
						}
					}
				},
				"description": "Verifies a user's email using a secure token.",
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/EmailVericationBodyInput"
							}
						}
					}
				}
			}
		},
...
		"/admin": {
			"post": {
				"operationId": "CreateAdmin",
				"responses": {
					"201": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIResponse_null_"
								}
							}
						}
					}
				},
				"description": "Creates a new admin user account with a given email and password.",
				"tags": [
					"Admin"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "The requst body object with zod validattion.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateUserInputBody",
								"description": "The requst body object with zod validattion."
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetAllUsers",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIResponse_PaginatedResult_UserWithRelationship_-or-null_"
								}
							}
						}
					}
				},
				"description": "Gets a paginated list of all general users with their settings.",
				"tags": [
					"Admin"
				],
				"security": [
					{
						"bearer": [
							"admin",
							"staff"
						]
					}
				],
				"parameters": [
					{
						"description": "Defaults to 10.",
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "What cursor field to work with.",
						"in": "query",
						"name": "nextCursor",
						"required": false,
						"schema": {
							"format": "date-time",
							"type": "string"
						}
					},
					{
						"description": "If the user's email is verified.",
						"in": "query",
						"name": "emailVerified",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					}
				]
			}
		},
		"/admin/organizers": {
			"get": {
				"operationId": "GetAllOrganizers",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIResponse_PaginatedResult_UserWithRelationship_-or-null_"
								}
							}
						}
					}
				},
				"description": "Gets a paginated list of all organizers with their related accounts and settings.",
				"tags": [
					"Admin"
				],
				"security": [
					{
						"bearer": [
							"admin",
							"staff"
						]
					}
				],
				"parameters": [
					{
						"description": "Defaults to 10.",
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "What cursor field to work with.",
						"in": "query",
						"name": "nextCursor",
						"required": false,
						"schema": {
							"format": "date-time",
							"type": "string"
						}
					},
					{
						"description": "If the user's email is verified.",
						"in": "query",
						"name": "emailVerified",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					}
				]
			}
		},
		"/admin/staffs": {
			"get": {
				"operationId": "GetAllStaffs",
				"responses": {
					"200": {
						"description": "Fetched all staffs",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIResponse_PaginatedResult_UserWithRelationship_-or-null_"
								}
							}
						}
					}
				},
				"description": "Gets a paginated list of all staffs with their related accounts and settings.",
				"tags": [
					"Admin"
				],
				"security": [
					{
						"bearer": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"description": "Defaults to 10.",
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "What cursor field to work with.",
						"in": "query",
						"name": "nextCursor",
						"required": false,
						"schema": {
							"format": "date-time",
							"type": "string"
						}
					},
					{
						"description": "If the user's email is verified.",
						"in": "query",
						"name": "emailVerified",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/api/v1"
		}
	]
}